Moved SplineBorisSequence to _temp
